+++ pysat/_instrument.py
@@ -1084,8 +1084,8 @@ class Instrument(object):
                 return
             elif isinstance(key, str):
                 # Assigning basic variables
-                if isinstance(in_data, xr.DataArray):
-                    # If xarray input, take as is
+                if isinstance(in_data, (xr.DataArray, tuple)):
+                    # If xarray or tuple input, take as is
                     self.data[key] = in_data
                 elif len(np.shape(in_data)) <= 1:
                     # If not an xarray input, but still iterable, then we
@@ -1129,12 +1129,9 @@ class Instrument(object):
                 else:
                     # Multidimensional input that is not an xarray.  The user
                     # needs to provide everything that is required for success.
-                    if isinstance(in_data, tuple):
-                        self.data[key] = in_data
-                    else:
-                        raise ValueError(' '.join(('Must provide dimensions',
-                                                   'for xarray multidim',
-                                                   'data using input tuple.')))
+                    # Passes the data through to get appropriate error from
+                    # xarray.
+                    self.data[key] = in_data
 
             elif hasattr(key, '__iter__'):
                 # Multiple input strings (keys) are provided, but not in tuple
@@ -3206,12 +3203,22 @@ class Instrument(object):
                     self._next_data, self._next_meta = self._load_next()
 
             # Make sure datetime indices for all data is monotonic
+            if self.pandas_format:
+                sort_method = "sort_index"
+                sort_args = []
+            else:
+                sort_method = 'sortby'
+                sort_args = ['time']
+
             if not self._index(self._prev_data).is_monotonic_increasing:
-                self._prev_data.sort_index(inplace=True)
+                self._prev_data = getattr(self._prev_data,
+                                          sort_method)(*sort_args)
             if not self._index(self._curr_data).is_monotonic_increasing:
-                self._curr_data.sort_index(inplace=True)
+                self._curr_data = getattr(self._curr_data,
+                                          sort_method)(*sort_args)
             if not self._index(self._next_data).is_monotonic_increasing:
-                self._next_data.sort_index(inplace=True)
+                self._next_data = getattr(self._next_data,
+                                          sort_method)(*sort_args)
 
             # Make tracking indexes consistent with new loads
             if self._load_by_date: